Rorary encoder_MQTT.C++
We have an NB-IoT application that uses SIM7000E shield and a microcontroller atmega328p ->Arduino Uno. We are looking for any interrupt comes from rotary encoder, if we get detect interrupt (change into pin1 of rotary encoder) we are going into function called handleEncoderChange. Later in the loop we got 
if (encoderChanged) =TRUE so SIM7000 is turning on and connects to a MQTT broker to send a value came from rotary encoder